Redis: Combine all the source files redis/commands/* into redis/commands.{hh,cc}
Fixes: #5394
